Move validation to separate package
Note that this changes api compatibility for earlier 2.2 RCs
