+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/KlumValidationException.java
@@ -24,6 +24,7 @@
 package com.blackbuild.klum.ast.util;
 
 import com.blackbuild.groovy.configdsl.transform.Validate;
+import com.blackbuild.klum.ast.validation.KlumValidationResult;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -37,7 +38,9 @@ import static java.util.stream.Collectors.joining;
  *     Note that this is explicitly not a subclass of {@link KlumModelException}, since a KlumModelException
  *     points to a specific model element that is invalid, while this exception collects multiple validation errors.
  * </p>
+ * @deprecated use {@link com.blackbuild.klum.ast.validation.KlumValidationException} instead.
  */
+@Deprecated
 public class KlumValidationException extends KlumException {
 
     private final List<KlumValidationResult> validationResults;
