+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/validation/KlumValidationIssue.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.blackbuild.klum.ast.util;
+package com.blackbuild.klum.ast.validation;
 
 import com.blackbuild.groovy.configdsl.transform.Validate;
 import org.jetbrains.annotations.NotNull;
