+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/validation/ValidationPhase.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.blackbuild.klum.ast.util;
+package com.blackbuild.klum.ast.validation;
 
 import com.blackbuild.klum.ast.process.DefaultKlumPhase;
 import com.blackbuild.klum.ast.process.VisitingPhaseAction;
@@ -39,6 +39,6 @@ public class ValidationPhase extends VisitingPhaseAction {
 
     @Override
     protected void doVisit(@NotNull String path, @NotNull Object element, @Nullable Object container, @Nullable String nameOfFieldInContainer) {
-        new Validator(element).execute();
+        new ObjectValidator(element).execute();
     }
 }
