+++ klum-ast-runtime/src/main/resources/META-INF/services/com.blackbuild.klum.ast.process.PhaseAction
@@ -1,9 +1,9 @@
-com.blackbuild.klum.ast.util.EarlyValidationPhase
+com.blackbuild.klum.ast.validation.EarlyValidationPhase
 com.blackbuild.klum.ast.util.OwnerPhase
 com.blackbuild.klum.ast.util.DefaultPhase
 com.blackbuild.klum.ast.util.layer3.AutoCreationPhase
 com.blackbuild.klum.ast.util.layer3.AutoLinkPhase
 com.blackbuild.klum.ast.util.PostTreePhase
-com.blackbuild.klum.ast.util.ValidationPhase
+com.blackbuild.klum.ast.validation.ValidationPhase
 com.blackbuild.klum.ast.util.VerifyPhase
 com.blackbuild.klum.ast.util.CleanupPhase
