+++ klum-ast-runtime/src/test/groovy/com/blackbuild/klum/ast/util/ValidatorTest.groovy
@@ -24,6 +24,7 @@
 package com.blackbuild.klum.ast.util
 
 import com.blackbuild.groovy.configdsl.transform.Validate
+import com.blackbuild.klum.ast.validation.ObjectValidator
 import spock.lang.Issue
 
 @SuppressWarnings("GrPackage")
@@ -196,7 +197,7 @@ class ValidatorTest extends AbstractRuntimeTest {
     }
 
     private static void validate(Object instance) {
-        def validator = new Validator(instance)
+        def validator = new ObjectValidator(instance)
         validator.execute()
         validator.validationIssues.throwOn(Validate.Level.ERROR);
     }
